+++ backend/vkt/src/test/resources/payment/paytrail-request.json
@@ -1,1 +1,1 @@
-{"stamp":"1","reference":"1","amount":100,"currency":"EUR","language":"FI","items":[{"unitPrice":0,"units":0,"vatPercentage":0,"productCode":"foo"},{"unitPrice":0,"units":0,"vatPercentage":0,"productCode":"bar"}],"redirectUrls":{"successUrl":"http://localhost/sucess","cancelUrl":"http://localhost/cancel"},"customer":{"email":"testinen@test.invalid","phone":null,"firstName":null,"lastName":null}}
+{"stamp":"1","reference":"1","amount":100,"currency":"EUR","language":"FI","items":[{"unitPrice":0,"units":0,"vatPercentage":0,"productCode":"foo"},{"unitPrice":0,"units":0,"vatPercentage":0,"productCode":"bar"}],"redirectUrls":{"success":"http://localhost/sucess","cancel":"http://localhost/cancel"},"customer":{"email":"testinen@test.invalid","phone":null,"firstName":null,"lastName":null}}
